add actual js changes

+++ fetch.js
@@ -1,15 +1,37 @@
+'use strict';
+
+Object.defineProperty( String.prototype, 'toCheerioObject', {
+    value: function() {
+        return cheerio.load(this);
+    },
+});
+
 const fs       = require('fs'),
       path     = require('path'),
       fetch    = require('node-fetch'),
       cheerio  = require('cheerio'),
+      formatting = require('./format'),
       baseUrls = [
     { url: 'https://forums.spacebattles.com/threads/imperium-ascendant-heresy-less-40k.596194/reader/', pages: 6 },
     { url: 'https://forums.spacebattles.com/threads/imperium-ascendant-heresy-less-40k.596194/16/reader/', pages: 3 },
     { url: 'https://forums.spacebattles.com/threads/imperium-ascendant-heresy-less-40k.596194/13/reader/', pages: 2 },
     { url: 'https://forums.spacebattles.com/threads/imperium-ascendant-heresy-less-40k.596194/19/reader/', pages: 3 },
       ],
-     baseFileDir = '/pages-forum/';
-     
+     pagesForum = '/pages-forum/',
+     pipe = (...fns) => (x) => fns.reduce((v, f) => f(v), x);
+
+function buildPost(post, filePath) {
+    return pipe(
+            formatting.removeStyle,
+            formatting.addStorytextDiv,
+            formatting.addLocalCss,
+            formatting.addJQueryCDN,
+            formatting.addLocalJs,
+            formatting.addSidebarButtons(filePath),
+        )
+    (post);
+}
+
 let pages = (() => {
     let arr = [
         ...baseUrls.map(obj => obj.url),
@@ -48,11 +70,16 @@ Promise.all(pages).then(
                                     .replace(/"/g, '')
                                     .replace(/\.?New$/, ''),
                                 time = new Date($('time', this).attr('datetime')),
-                                post = $('div.bbWrapper', this).html();
-                            // console.log(path.join(__dirname, `/pages-forum/${type}_${title}_${time}.html`))
+                                fileName = `${pagesForum}${time.getTime()}--${type}_${title}.html`,
+                                filePath = path.join(__dirname, fileName),
+                                post = $('div.bbWrapper', this).html(),
+                                builtPost = buildPost(post, filePath);
+                            if (fs.existsSync(filePath) /* || forceRewrite */) {
+                                return;
+                            };
                             fs.writeFileSync(
-                                path.join(__dirname, `${baseFileDir}${time.getTime()}--${type}_${title}.html`),
-                                '<link rel="stylesheet" href="asc.css"><div class="storytext">' + post + '</div>'
+                                filePath,
+                                builtPost
                             );   
                         });
                     }
@@ -63,3 +90,4 @@ Promise.all(pages).then(
     },
     reject => console.log(reject).catch(err => console.log(err))
 );
+
